requirements-deps.txt
click 
pydantic 
loguru 
pyyaml 
psutil
rabbitmq
eventlet
kombu
werkzeug